soporte para BigDecimals

+++ src/test/java/evaluator/EvaluatorTest.java
@@ -3,6 +3,7 @@ package evaluator;
 import org.junit.jupiter.api.Test;
 import parser.Parser;
 
+import java.math.BigDecimal;
 import java.util.Arrays;
 import java.util.List;
 
@@ -17,7 +18,7 @@ class EvaluatorTest {
                                         ("-7", "-3", "-3.33", "4", "*", "+", "-20", "10", "/", "-", "2", "^", "/");
         double expResult = -0.034135950813020816;
 
-        double result = Evaluator.evaluateExpression(psofixExpression);
+        BigDecimal result = Evaluator.evaluateExpression(psofixExpression);
         assertEquals(expResult, result);
     }
 
@@ -27,7 +28,7 @@ class EvaluatorTest {
                 ("0", "sin", "-5.5", "-2", "abs", "*", "-");
         double expResult = 11;
 
-        double result = Evaluator.evaluateExpression(psofixExpression);
+        BigDecimal result = Evaluator.evaluateExpression(psofixExpression);
         assertEquals(expResult, result);
     }
 
@@ -37,7 +38,7 @@ class EvaluatorTest {
                 ("-10.5", "2", "2", "^", "-3", "1", "+", "-", "/", "2", "*");
         double expResult = -3.5;
 
-        double result = Evaluator.evaluateExpression(psofixExpression);
+        BigDecimal result = Evaluator.evaluateExpression(psofixExpression);
         assertEquals(expResult, result);
     }
 
@@ -47,7 +48,7 @@ class EvaluatorTest {
                 ("100", "log10", "3.5", "-1.5", "+", "2", "^", "*");
         double expResult = 400;
 
-        double result = Evaluator.evaluateExpression(psofixExpression);
+        BigDecimal result = Evaluator.evaluateExpression(psofixExpression);
         assertEquals(expResult, result);
     }
 
@@ -57,7 +58,7 @@ class EvaluatorTest {
                 ("2", "3", "2", "^", "^", "NEG", "500", "12", "+", "/", "1.5", "+");
         double expResult = 2.5;
 
-        double result = Evaluator.evaluateExpression(psofixExpression);
+        BigDecimal result = Evaluator.evaluateExpression(psofixExpression);
         assertEquals(expResult, result);
     }
